+++ cmd/agent/main.go
@@ -18,20 +18,15 @@ import (
 	"time"
 )
 
-// Используйте go run -ldflags "-X main.buildVersion=v1.0.1 -X main.buildDate=11.11.2024 -X main.buildCommit=caferacer"
-// Чтобы установить переменные для сборки
 var (
-	valuesGauge  = map[string]float64{}
-	pollCount    uint64
-	mu           sync.Mutex
-	buildVersion = "N/A"
-	buildDate    = "N/A"
-	buildCommit  = "N/A"
+	valuesGauge = map[string]float64{}
+	pollCount   uint64
+	mu          sync.Mutex
 )
 
 func main() {
 
-	printBuildInfo()
+	config.PrintBuildInfo()
 	cfg := config.NewClient()
 	var wg sync.WaitGroup
 
@@ -71,19 +66,6 @@ func main() {
 		}
 	}()
 	wg.Wait()
-}
-
-func printBuildInfo() {
-	printOrPanic(fmt.Sprintf("Build version: %v", buildVersion))
-	printOrPanic(fmt.Sprintf("Build date: %v", buildDate))
-	printOrPanic(fmt.Sprintf("Build commit: %v", buildCommit))
-}
-
-func printOrPanic(data string) {
-	_, err := fmt.Println(data)
-	if err != nil {
-		panic(err)
-	}
 }
 
 func getMetrics() {
